+++ megaparsec.cabal
@@ -32,9 +32,12 @@ flag dev
   manual:             True
   default:            False
 
+flag with-base-earlier-than-4-13
+  description:        build with base earlier than 4.13.
+  default:            False
+
 library
-  build-depends:      base         >= 4.11  && < 5.0
-                    , bytestring   >= 0.2   && < 0.11
+  build-depends:      bytestring   >= 0.2   && < 0.11
                     , case-insensitive >= 1.2 && < 1.3
                     , containers   >= 0.5   && < 0.7
                     , deepseq      >= 1.3   && < 1.5
@@ -52,6 +55,8 @@ library
                     , Text.Megaparsec.Error
                     , Text.Megaparsec.Error.Builder
                     , Text.Megaparsec.Internal
+                    , Text.Megaparsec.Internal.Monad
+                    , Text.Megaparsec.Internal.ParsecT
                     , Text.Megaparsec.Pos
                     , Text.Megaparsec.Stream
   other-modules:      Text.Megaparsec.Class
@@ -67,6 +72,12 @@ library
                       -Wincomplete-record-updates
                       -Wincomplete-uni-patterns
                       -Wnoncanonical-monad-instances
+  if flag(with-base-earlier-than-4-13)
+    build-depends: base >= 4.11 && < 4.13
+    hs-source-dirs:   . base-earlier-than-4-13
+  else
+    build-depends: base >= 4.13 && < 5.0
+    hs-source-dirs:   . base-4-13-or-later
   default-language:   Haskell2010
 
 benchmark bench-speed
